requirements.txt
astroid==1.6.3
autopep8==1.3.5
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.6
isort==4.3.4
jieba3k==0.35.1
lazy-object-proxy==1.3.1
lxml==4.2.1
mccabe==0.6.1
mysqlclient==1.3.12
newspaper3k==0.2.6
nltk==3.3
numpy==1.14.3
Pillow==5.1.0
pkg-resources==0.0.0
pycodestyle==2.4.0
pylint==1.8.4
python-dateutil==2.7.2
PyYAML==3.12
requests==2.18.4
requests-file==1.4.3
six==1.11.0
tldextract==2.2.0
urllib3==1.22
wrapt==1.10.11
